Moves lifecycle callbacks to their own file and makes Dispatchers a data class
