+++ firebase-sessions/src/main/kotlin/com/google/firebase/sessions/Dispatchers.kt
@@ -21,10 +21,10 @@ import com.google.firebase.app
 import kotlin.coroutines.CoroutineContext
 
 /** Container for injecting dispatchers. */
-class Dispatchers
-internal constructor(
-  internal val blockingDispatcher: CoroutineContext,
-  internal val backgroundDispatcher: CoroutineContext,
+internal data class Dispatchers
+constructor(
+  val blockingDispatcher: CoroutineContext,
+  val backgroundDispatcher: CoroutineContext,
 ) {
 
   companion object {
